# Swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Sistema de Subastas API",
      "description": "API REST del Sistema de Subastas — UADE Desarrollo de Aplicaciones I, 1C 2026. Gestiona registración de postores, catálogo, subastas, pujas, ventas, multas, seguros y solicitudes de incorporación de bienes.",
      "contact": {
        "name": "Equipo DA1",
        "email": "agustin.torres@majily.com"
      },
      "version": "0.1.0"
    },
    "paths": {
...
      "/ventas/cerrar/{catalogo_item_id}": {
        "post": {
          "tags": [
            "Ventas"
          ],
          "summary": "Cerrar la venta tomando la mejor puja del ítem",
          "operationId": "cerrar_venta_ventas_cerrar__catalogo_item_id__post",
          "parameters": [
            {
              "name": "catalogo_item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Catalogo Item Id"
              }
            },
            {
              "name": "medio_pago_id",
              "in": "query",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Medio Pago Id"
              }
            },
            {
              "name": "comision_pct",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ],
                "default": "0.10",
                "title": "Comision Pct"
              }
            },
            {
              "name": "costo_envio",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ],
                "default": "0",
                "title": "Costo Envio"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/VentaOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/ventas/{venta_id}/impago": {
        "post": {
          "tags": [
            "Ventas"
          ],
          "summary": "Registrar impago: genera multa del 10% y bloquea al usuario",
          "operationId": "registrar_impago_ventas__venta_id__impago_post",
          "parameters": [
            {
              "name": "venta_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Venta Id"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/ventas/{venta_id}": {
        "get": {
          "tags": [
            "Ventas"
          ],
          "summary": "Obtener Venta",
          "operationId": "obtener_venta_ventas__venta_id__get",
          "parameters": [
            {
              "name": "venta_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Venta Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/VentaOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/ventas/{venta_id}/pagar": {
        "post": {
          "tags": [
            "Ventas"
          ],
          "summary": "Marcar venta como pagada",
          "operationId": "marcar_pagada_ventas__venta_id__pagar_post",
          "parameters": [
            {
              "name": "venta_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Venta Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/VentaOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
        "VentaOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "catalogo_item_id": {
              "type": "integer",
              "title": "Catalogo Item Id"
            },
            "comprador_id": {
              "type": "integer",
              "title": "Comprador Id"
            },
            "medio_pago_id": {
              "type": "integer",
              "title": "Medio Pago Id"
            },
            "monto_final": {
              "type": "string",
              "title": "Monto Final"
            },
            "comision": {
              "type": "string",
              "title": "Comision"
            },
            "costo_envio": {
              "type": "string",
              "title": "Costo Envio"
            },
            "retira_personalmente": {
              "type": "boolean",
              "title": "Retira Personalmente"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda"
            },
            "fecha": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha"
            },
            "pagada": {
              "type": "boolean",
              "title": "Pagada"
            }
          },
          "type": "object",
          "required": [
            "id",
            "catalogo_item_id",
            "comprador_id",
            "medio_pago_id",
            "monto_final",
            "comision",
            "costo_envio",
            "retira_personalmente",
            "moneda",
            "fecha",
            "pagada"
          ],
          "title": "VentaOut"
        }
      }
    }
  }